Moved helper class to helpers directory to be ignored by ava
