+++ test/unit/public-sdk-apis/sendUniqueOutcome.ts
@@ -10,7 +10,7 @@ import MainHelper from "../../../src/helpers/MainHelper";
 import Log from "../../../src/libraries/Log";
 import Database from "../../../src/services/Database";
 import timemachine from "timemachine";
-import OutcomeTestHelper from '../helpers1/OutcomeTestHelper';
+import OutcomeTestHelper from '../helpers/OutcomeTestHelper';
 import { SentUniqueOutcome } from '../../../src/models/Outcomes';
 import { isArrayContentsEqual } from '../../support/tester/utils';
